trainer/requirements.txt
xgboost
numpy
pandas
scikit-learn
psycopg2-binary
